moved ToggleProps to types folder
