+++ components/tools/Toggle.tsx
@@ -1,23 +1,6 @@
 import { twMerge } from 'tailwind-merge';
 
-export interface ToggleProps {
-  // eslint-disable-next-line prettier/prettier
-
-  /** Current state of the toggle. */
-  checked: boolean;
-
-  /** Function to update the toggle state. */
-  setChecked: React.Dispatch<React.SetStateAction<boolean>>;
-
-  /** Label text for the toggle. */
-  label?: string;
-
-  /** The background color of the checkbox. */
-  bgColor?: string;
-
-  /** The background color of the checkbox when it is checked. */
-  checkedStateBgColor?: string;
-}
+import type { ToggleProps } from '@/types/components/tools/TogglePropsType';
 
 /**
  * Toggle component for displaying and controlling a toggle switch.
